Pruebas y Configuración de Seeders en Proyectos Laravel.

+++ database/migrations/2026_04_21_155407_create_repositories_table.php
@@ -13,7 +13,8 @@ return new class extends Migration
     {
         Schema::create('repositories', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
+            // $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users'); // Make de connect between repositories to users.
             $table->string('url');
             $table->text('description');
